+++ src/main/java/com/github/celestial_awakening/capabilities/MoonScytheCapability.java
@@ -8,19 +8,19 @@ Crit hits create a crescent strike
 Non-crit hits create a crescent wave
 Each ability has a cool down of 5 seconds.
      */
-    int strikeCD=0;
-    int waveCD=0;
-    int lunarOrbs=0;
+    short strikeCD=0;
+    short waveCD=0;
+    byte lunarOrbs=0;
     int lastOrbTick=0;
     void saveNBTData(CompoundTag nbt){
-        nbt.putInt("StrikeCD",strikeCD);
-        nbt.putInt("WaveCD",waveCD);
-        nbt.putInt("LunarOrbs",lunarOrbs);
+        nbt.putShort("StrikeCD", (short) strikeCD);
+        nbt.putShort("WaveCD", (short) waveCD);
+        nbt.putByte("LunarOrbs", (byte) lunarOrbs);
     }
     void loadNBTData(CompoundTag nbt){
-        strikeCD=nbt.getInt("StrikeCD");
-        waveCD=nbt.getInt("WaveCD");
-        lunarOrbs=nbt.getInt("LunarOrbs");
+        strikeCD=nbt.getShort("StrikeCD");
+        waveCD=nbt.getShort("WaveCD");
+        lunarOrbs=nbt.getByte("LunarOrbs");
     }
     public int getStrikeCD(){
         return strikeCD;
@@ -38,15 +38,15 @@ Each ability has a cool down of 5 seconds.
         strikeCD+=cnt;
     }
     public void changeLunarOrbs(int cnt,int t){
-        lunarOrbs=Math.max(0,Math.min(6,cnt));
+        lunarOrbs= (byte) Math.max(0,Math.min(6,cnt));
         lastOrbTick=t;
     }
     public void incrementLunarOrbs(int t){
-        lunarOrbs=Math.min(6,lunarOrbs+1);
+        lunarOrbs= (byte) Math.min(6,lunarOrbs+1);
         lastOrbTick=t;
     }
     public void decrementLunarOrbs(int t){
-        lunarOrbs=Math.max(0,lunarOrbs-1);
+        lunarOrbs= (byte) Math.max(0,lunarOrbs-1);
         lastOrbTick=t;
     }
     public int getLastOrbTick(){
