+++ src/main/java/com/github/celestial_awakening/capabilities/SunStaffCapability.java
@@ -19,24 +19,24 @@ While holding shift and right click, create an expanding zone that deals fire da
 The zone can last for up to 10 seconds, and the area can have a max radius of 7 blocks.
 CD of 4-10 seconds, depending on how long the storm lasted.
  */
-    int flashCD=0;
-    int rayCD=0;
-    int windCD=0;
+    short flashCD=0;
+    short rayCD=0;
+    short windCD=0;
     void saveNBTData(CompoundTag nbt){
-        nbt.putInt("FlashCD",flashCD);
-        nbt.putInt("RayCD",rayCD);
-        nbt.putInt("WindCD",windCD);
+        nbt.putShort("FlashCD",flashCD);
+        nbt.putShort("RayCD",rayCD);
+        nbt.putShort("WindCD",windCD);
     }
     void loadNBTData(CompoundTag nbt){
-        flashCD=nbt.getInt("FlashCD");
-        rayCD=nbt.getInt("RayCD");
-        windCD=nbt.getInt("WindCD");
+        flashCD=nbt.getShort("FlashCD");
+        rayCD=nbt.getShort("RayCD");
+        windCD=nbt.getShort("WindCD");
     }
 
     public void decrementCD(){
-        flashCD=Math.max(0,flashCD-1);
-        rayCD=Math.max(0,rayCD-1);
-        windCD=Math.max(0,windCD-1);
+        flashCD= (short) Math.max(0,flashCD-1);
+        rayCD= (short) Math.max(0,rayCD-1);
+        windCD= (short) Math.max(0,windCD-1);
     }
 
     public int getFlashCD(){
@@ -52,14 +52,14 @@ CD of 4-10 seconds, depending on how long the storm lasted.
     }
 
     public void setFlashCD(int flashCD) {
-        this.flashCD = Math.max(0,flashCD);
+        this.flashCD = (short) Math.max(0,flashCD);
     }
 
     public void setRayCD(int rayCD) {
-        this.rayCD = Math.max(0,rayCD);
+        this.rayCD = (short) Math.max(0,rayCD);
     }
 
     public void setWindCD(int windCD) {
-        this.windCD = Math.max(0,windCD);
+        this.windCD = (short) Math.max(0,windCD);
     }
 }
